Separate out image snapshots by devicePixelRatio
